requirements.txt
alembic==1.11.3
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.3
fastapi==0.101.1
fastapi-utils==0.2.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mergepythonclient==0.1.1
orjson==3.9.5
psycopg2==2.9.7
psycopg2-binary==2.9.7
pydantic==1.10.12
pydantic-extra-types==2.0.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
typing-extensions==4.7.1
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
zipp==3.16.2
